Move Get Started to own panel component
